So now blend is mixin.
The mixin method mixes all passed objects except the first *into* the
first object. So you pass the prototype of a class first and then all of
your mixin objects after. It will mix all of the methods and values into
the prototype/object you passed as the first argument.