Now escaping routes. No RegExp injection.
Also fixed a few little RegExp related routing issues. Now the only
valid route segment is \w, so that's a-z, 0-9 and underscores. So you
can't put hyphens in route segments anymore. It would not have worked
with the escaping.

But you can now use hyphens in the actual URL / slug / request.