Removed find, for now.
Does not actually apply to the structure of the default object storage
method. You should fetch your objects and filter them in some other way
really.

If you had a server endpoint you could add a find method that would
actually ask the server with a search term that the backend can parse.
It just doesn't apply to the browser and objects correctly.